add consultar.py to legados + absolute imports
