+++ backend/pgdas_fiscal_oesk/export2same_folder.py
@@ -1,7 +1,7 @@
 # dale
 from re import search
 import openpyxl
-from contimatic import Contimatic
+from pgdas_fiscal_oesk.contimatic import Contimatic
 import os
 from shutil import copy2
